refactor(terminal): split input handling
